proj/rag/pyproject.toml
[tool.poetry]
name = "twiml-rag"
version = "0.1.0"
description = "twiml-rag application & model configuration (+ exploratory jupyter-nbs)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    {include="rag"},
]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
streamlit-chat = "^0.1.1"
langchain = "^0.0.350"
sentence-transformers = "^2.2.2"
python-dotenv = "^1.0.0"
unstructured = {extras = ["md"], version = "^0.11.2"}
openai = "^1.3.9"
pinecone-client = "^2.2.4"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]] # ignoring missing stubs for pkgs
module = [
    "sentence_transformers",
    "streamlit_chat",
    "pinecone"
]
ignore_missing_imports = true
